Add simple server for autotesting
Since running the entire build.R script is pretty slow due to
loading the R dependencies, I added a function to the script that
runs a socket server. When an incoming connection is established,
it runs the test code. This provides a much faster turnaround
since the libraries only have to be loaded once when the script
starts.